core/server/requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.63.0
h11==0.13.0
idna==3.3
motor==2.2.0
numpy==1.20.0
nvdlib==0.5.6
passlib==1.7.4
progressbar==2.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
pymongo==3.12.3
pyparsing==3.0.7
pysmile==1.6.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
six==1.16.0
starlette==0.13.6
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.13.4
xmltodict==0.13.0
